MVCF: Avoid caching stale verbs in VerbManager
MVCF currently patches OrderForceAttack() and TryGetAttackVerb() if any
of its features were activated by a mod in the modlist; these will then
affect verb choice for any pawn. However, this can pose a problem if the
current modlist does not activate Feature_Humanoid, or if Combat
Extended is active: it can lead to a situation where VerbManager caches
a stale CurrentVerb forever, and use it even after the pawn has changed
weapons.

A simple reproduction follows:
(1) Have a mod in the modlist that uses some MVCF features - e.g. vanilla animals caves
(2) Start quicktest map
(3) Equip starting rifle, draft colonist, shoot at some crap
(4) Equip starting knife
(5) Spawn manhunter
(6) Colonist will shoot with the knife

As a fix, move EquipmentTracker-related patches from Feature_Humanoid to
Feature, so that pawns' equipment changes are correctly tracked if MVCF
is active. For this to work correctly, also patch
EquipmentTracker.TryTransferEquipmentToContainer, which is called in
some vanilla edge cases (e.g. when a pawn is downed while onboard a
transport pod or shuttle), but is also used to implement Combat
Extended's inventory system. Finally, update VerbManager.RemoveVerb() to
null out its cached CurrentVerb if it was the verb that just got
removed.

On existing saves, it's possible for error spam to occur if a pawn
managed to cache a ranged verb inside its VerbManager, then later
switched to a melee weapon - when this pawn is drafted, it'll attempt to
use its cached, discarded ranged verb to "scan" for targets via
CheckForAutoAttack() as if they were wielding a ranged weapon, and cause
errors as the `caster` for this verb will be `null` if it was discarded.
Avoid this by adding a basic logic check to VerbManager to null out the
cached CurrentVerb value on init if it's no longer wielded by this pawn.