Added a number of rendering and size related functions for scaling pawns for the VRE Pigskin projects where we want "Ever Growing" pigs with a size that gets adjusted based on their biological age. + A lot of support code ported and rewritten from my past work on Big & Small so we can deal with other 'on the fly' behaviour, e.g. smaller pawns. + Support for modifying the food-bar-size.
The current rendering effects of changed size are:
* Enlarged pawns get offset upwards so their 'feet' are still on the ground.
* Enlarged pawns get smaller heads relative to their body (capped)
* Enlarged pawns get exponentially less large to stick to the Rimworld style.
* Shrunken pawns get larger "chibi" heads much like children or small animals in Rimworld.

The current gameplay effects of size are:
* Increased Health.
* Increased BodySize
This does not include other stuff from B&S, such as increased hunger, somewhat increased damage, increased armour pen, reduced dodge. (and inverted for small pawns). If any of this is wished for it can be integrated and rewritten though.

Since there is a lot of Math and StatDef lookups a generics-based caching system was made. This can sometimes look a bit strange in the profiler since a very random method tends to take all the blame for all other method's calls.