Swapped over to using DefOfs instead of calling "Named", as discussed.
Fixed BodySize being overwritten instead of added to for better compatibility with other mods that change it in the same way (e.g. Big & Small)

Improved on the health calculations, a Size 2 pawn will now have almost exactly the health of a bear (~also around Size 2), while a Size 4 pawn will be in the ballpark of a Thrumbo.

Absolutely tiny change to handle entities (e.g. animals) that don't have a AgeBiologicalYearsFloat. Not an issue currently since it only runs on humans, but I figure handling the Null is still good in case someone wants to scale animals later.