Fix KCSG required mod ID check
Basically, `PackageId` could return the mod with a `_steam` postfix if running a Steam release with a local mod installed. Since the mod ID in XML wouldn't include it, it would cause issues.

The simplest fix is to use `PackageIdNonUnique` since it is guaranteed to not include the `_steam` postfix.