Update CompAmbientSound
I've changed how the comp determines if a sustainer should play or not. Using only signals didn't work too well, since if it received one signal, the sound would always play - ignoring any other conditions. So if an object has no power but is flicked on, the sustainer would restart since it received "flicked on" signal.

List of changes:
- Additional check if every possible comp is either null or enabled before (re)starting the sustainer
  - All the comps have also been cached, so there's no need to use `GetComp` each time a change happens
- Added support to `CompRefuelable`, so the sound will only play if a building is fueled
- Removed support for `CompAutoPowered`
  - Because of the way this comp works, we would need to use ticking to check if the sustainer can turn on
  - If support for this comp is required for any reason, let me know, and I'll re-add the support for it
- A couple of other minor changes to the code