Add comp for customizable graphics
The comp itself requires the Thing to use either a subtype of `Graphic_Indexed`, or `Graphic_Random` (only tested with indexed only).

The comp provides a way to freely switch between those graphics (with a preview when using the float menu), allows selecting which graphic should be the default one. It should also fully support styled buildings.

Due to some Harmony patches, I've made sure it's possible to change the graphic of frames and blueprints (but not install blueprints). However, when using install blueprints, changing the original building will also update the blueprint.

For now, I haven't included the ability to modify a minified building, as I've decided it's time to move on to other tasks. I may return to this at some point again and implement it, but not right now.