Add a mod extension for linkables to link only on interaction spots
A new `DefModExtension` called `FacilityExtension`. At the moment, it only has a single field (`linkOnInteractionSpots`) which, if true, will only allow if their building overlaps any of the interaction spots.

There was a need for a Harmony patch, which checks if the def contains `FacilityExtension` with `linkOnInteractionSpots` being true. It'll then check if the linkable overlaps any of the interaction spots (skipping any of the vanilla linkable checks).

The Harmony patch will only be applied if there's at least a single def that has `FacilityExtension` with `linkOnInteractionSpots` to reduce the amount Harmony patches that end up being unused.