Added CompScheduleActiveAtSunlightRange
It extends the vanilla CompSchedule (and should be compatible with stuff that checks for it, like `CompAmbientSound`).

On top of that, it should work with either of ticker types (but normal ticker will still re-check every 250 ticks like rare ticker).

The method `ShouldBeAllowed` is virtual, so other types extending this one can add extra checks if needed for anything.