Rename CompScheduleActiveAtSunlightRange to CompScheduleExtended, extend its functionality
CompScheduleExtended now supports disabling the building if it's (not) under a roof. It also has separate messages that can be setup when disabled due to light level and (not) being under a roof.

The sunlight check was changed, so the lamp should now actually work during an eclipse/unnatural darkness or other events that turn lower the map light level.

I've kept CompProperties_ScheduleActiveAtSunlightRange for now for compatibility reasons. Suddenly changing it would break Vanilla Furniture Expanded, so we'd need to update both at the same time (and hope Steam updates it at the same time for everyone). It's currently a subtype of CompProperties_ScheduleExtended and will be removed in the future.