Fix IncreaseRadialPatternRadiiSize issues
The `IncreaseRadialPatternRadiiSize` was called from `StaticConstructorOnStartup`, which was done too late. It ended up with `GenRadial` methods being inlined, optimized, etc. which caused issues. Calling it in a constructor for a `Mod` subclass seems to be early enough, so I've moved it to `VFEGlobal` class. If someone wants to move it somewhere else, feel free to do so.

As an additional side note, I've added a small condition to disable the radial pattern size increase if the current size is already the same or bigger (if, for example, a different mod applies it).