Fixed ApparelDrawPosExtension.secondaryApparelGraphics patch
The old patch could not work anymore as it used to before, as `ProcessApparel` method no longer calls `AddChild` on the render node. Instead, it returns a list of render nodes that will be added.

The way the patch works now is by populating an external list of render node pairs (node, parent), while handling the layer offsets for the parent. It will then add them in the postfix to the resulting IEnumerable from `ProcessApparel`.