Add a way to get a list of things spawned by KCSG
You may pass a `ICollection<Thing>` (so a List or a HashSet, for example), which will be filled with a list of spawned thigns/pawns/items/etc. - you may also pass null.

This doesn't handle shuttles, or mortars, shells, and pawns manning them. Those are handled differently and I don't see an easy way to hahndle it.